a.	If a user creates a new account and at least one field for the Billing address is not empty, the system must automatically create a new Main Billing Address. The same logic should be applied for Shipping Address (30 min) b.	One account can have only one main billing address and only one main shipping address (30 min) c.	If a user updates Account Billing Address field changes should be automatically applied to Main Billing Address. Same logic for Shipping Address (45 min)
d.	If a user updates Main Billing Address changes should be automatically applied to Account Billing Address. The same logic should be applied for Main Shipping Address (45 min)
e.
f.	Users should be able to create a new billing/shipping address and mark it as main. (40 min)
g.	On the standard related list component on an account object, display details of the account address, such that a user can see if it is a main shipping/billing address. (60 min)
h.	Customize a list view for account addresses such that we can see up to 10 fields on the ‘recently viewed list view’ and we should see the same 10 fields on the view all list (30 min)
i.	Create a list view that will show account addresses created today (30 min)
j.	For all main addresses that are not the same as the parent account, Mass update the addresses that are not the same as the parent account. Precise which solution will used when working with 50,000-250,000 records and 250,000-300,000 records. All solutions should be implemented. (120 min)
k.	Create a custom metadata that will switch off apex triggers (60 min)